moved grpc clients out of grpc function test
